Fixed a bug in sp_plot.c, adding labels to the time series plots. Fixed a bug in figuring out what goes in the cascading menus for   launching new displays -- edge sets were somehow being added still.

+++ splot.c
@@ -936,7 +936,7 @@ splot_cursor_set (gint jcursor, splotd *sp)
 {
   GdkWindow *window = sp->da->window;
 
-  if (jcursor == NULL) {
+  if (jcursor == (gint) NULL) {
     gdk_cursor_destroy (sp->cursor);
     sp->jcursor = (gint) NULL;
   } else {
